Day 9
The background change doesn’t work, the red circles move really weirdly and when they collide with the bomb they will relocate once, but then not again. I got the background change to work!! Now I need to find an actual night background and decide if I want the game to have a time limit, and if so decide at what time to change it to night. I also resized the red circles and changed the orientation of the bomb image.