Day 12
I got the end screen to partially work, but am still facing some issues as the text that I want to appear doesn’t. I looked at Red class and my Bomb class and began comparing/analyzing the two. Only one bomb can drop from the plane and be on the screen at a time, but there are multiple Red circles. I was wondering if I should change this. I was also thinking of trying to add tiles to my game, but I’d need to come up with how to incorporate them into my game.