Day 13
I tried to find an explosion image with a transparent background. I couldn’t find one for some reason, so for now there’s only a placeholder. I realized the plane image doesn’t change when it moves to the right, I fixed that but now there’s an issue with the scaling. There are now issue’s when I click the b key to drop the bomb, and the explosion image is not appearing. Though I did write some steps in my code on how it should work.
#a) when collide get the bomb to stop moving
#b)explosion image will replace bomb image
#c)appear/disapear quickly